+++ sweat/__init__.py
@@ -1,0 +1,1 @@
+from .examples.utils import examples, FileTypeEnum, SportEnum